Tenant modals validation

+++ modules/tenant-management/src/Volo.Abp.TenantManagement.Blazor/Pages/TenantManagement/TenantManagement.razor.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using Blazorise;
 using Microsoft.AspNetCore.Authorization;
@@ -20,9 +21,10 @@ namespace Volo.Abp.TenantManagement.Blazor.Pages.TenantManagement
         protected FeatureManagementModal FeatureManagementModal;
 
         protected Modal ManageConnectionStringModal;
+        protected Validations ManageConnectionStringValidations;
 
         protected TenantInfoModel TenantInfo;
-        
+
         public TenantManagement()
         {
             LocalizationResource = typeof(AbpTenantManagementResource);
@@ -36,7 +38,7 @@ namespace Volo.Abp.TenantManagement.Blazor.Pages.TenantManagement
 
             TenantInfo = new TenantInfoModel();
         }
-        
+
         protected async override Task SetPermissionsAsync()
         {
             await base.SetPermissionsAsync();
@@ -47,6 +49,8 @@ namespace Volo.Abp.TenantManagement.Blazor.Pages.TenantManagement
 
         protected virtual async Task OpenEditConnectionStringModalAsync(Guid id)
         {
+            ManageConnectionStringValidations.ClearAll();
+
             var tenantConnectionString = await AppService.GetDefaultConnectionStringAsync(id);
 
             TenantInfo = new TenantInfoModel
@@ -67,18 +71,21 @@ namespace Volo.Abp.TenantManagement.Blazor.Pages.TenantManagement
 
         protected virtual async Task UpdateConnectionStringAsync()
         {
-            await CheckPolicyAsync(ManageConnectionStringsPolicyName);
-
-            if (TenantInfo.UseSharedDatabase || TenantInfo.DefaultConnectionString.IsNullOrWhiteSpace())
+            if (ManageConnectionStringValidations.ValidateAll())
             {
-                await AppService.DeleteDefaultConnectionStringAsync(TenantInfo.Id);
-            }
-            else
-            {
-                await AppService.UpdateDefaultConnectionStringAsync(TenantInfo.Id, TenantInfo.DefaultConnectionString);
-            }
+                await CheckPolicyAsync(ManageConnectionStringsPolicyName);
 
-            ManageConnectionStringModal.Hide();
+                if (TenantInfo.UseSharedDatabase || TenantInfo.DefaultConnectionString.IsNullOrWhiteSpace())
+                {
+                    await AppService.DeleteDefaultConnectionStringAsync(TenantInfo.Id);
+                }
+                else
+                {
+                    await AppService.UpdateDefaultConnectionStringAsync(TenantInfo.Id, TenantInfo.DefaultConnectionString);
+                }
+
+                ManageConnectionStringModal.Hide();
+            }
         }
 
         protected override string GetDeleteConfirmationMessage(TenantDto entity)
@@ -93,6 +100,7 @@ namespace Volo.Abp.TenantManagement.Blazor.Pages.TenantManagement
 
         public bool UseSharedDatabase { get; set; }
 
+        [Required]
         public string DefaultConnectionString { get; set; }
     }
 }
